Split out common code into shared directory
