+++ keyboards/nullbitsco/common/bitc_led.c
@@ -1,4 +1,4 @@
-/* Copyright 2020 Jay Greco
+/* Copyright 2021 Jay Greco
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -13,17 +13,25 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#pragma once
+#include "bitc_led.h"
 
-#include "quantum.h"
+void set_bitc_LED(uint8_t mode) {
+    switch(mode) {
+        case LED_ON:
+            setPinOutput(PIN_LED);
+            writePin(PIN_LED, GPIO_STATE_HIGH);
+        break;
 
-#define LED_ON          2
-#define LED_DIM         1
-#define LED_OFF         0
+        case LED_DIM:
+            setPinInput(PIN_LED);
+        break;
 
-#define GPIO_STATE_LOW  0
-#define GPIO_STATE_HIGH 1
+        case LED_OFF:
+            setPinOutput(PIN_LED);
+            writePin(PIN_LED, GPIO_STATE_LOW);
+        break;
 
-#define PIN_LED         F0
-
-void set_bitc_LED(uint8_t mode);
+        default:
+        break;
+    }
+}
