+++ keyboards/nullbitsco/nibble/nibble.c
@@ -1,4 +1,4 @@
-/* Copyright 2020 Jay Greco
+/* Copyright 2021 Jay Greco
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -14,7 +14,6 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 #include QMK_KEYBOARD_H
-#include "bitc_led.h"
 
 // Use Bit-C LED to show CAPS LOCK status
 bool led_update_kb(led_t led_state) {
