+++ keyboards/nullbitsco/nibble/rules.mk
@@ -24,7 +24,7 @@ CUSTOM_MATRIX = lite        # Lite custom matrix
 
 # Project specific files
 SRC += matrix.c \
-       bitc_led.c \
+       common/bitc_led.c \
        big_led.c \
-       remote_kb.c
+       common/remote_kb.c
 QUANTUM_LIB_SRC += uart.c
